Create Master V2 Code
This code adds a light level counter that the main if statement  runs off of.

Also changed the ultrasonic reader to run off of the determineAndSendObject function

Also added the ultrasonic debug mode that prints readings to the screen. Device enters this mode when the A button is pressed and released.