Random Color Picker
Added a small function to choose random colors so single-color functions should no longer default to red. Made major changes to comments in the code for ease of understanding and to clean the code up.  Removed variables that weren't referenced.  Commented out Serial.print lines for production-ready intent. Reduced raw values from 0 to 700 to 0 to 650, since the ultrasonic could be triggered by a car in the immediate parking space.